+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 - For redis storage `json` encoder/decoder replaced with [msgpack](https://github.com/vmihailenco/msgpack)
 - Storage now accepts and returns request content as a bytes slice
+- Response content for session creation handler must be base64-encoded
+- All API handlers return request body and session response base64-encoded (JSON request and response property `content` replaced with `content_base64`)
+
 ## v0.3.1
 
 ### Changed